# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SynCloud API",
    "version": "v1",
    "description": "The public SynCloud controller API. The dashboard, synctl and CI all use these endpoints (plan \u00a75.1). Authenticate with the dashboard session cookie, a personal access token (Bearer), or a signed access-key request (SYN1-HMAC-SHA256, plan \u00a77.1)."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "session": []
    },
    {
      "bearerToken": []
    },
    {
      "accessKey": []
    }
  ],
  "paths": {
...
    "/api/v1/projects/{project}/environments/{env}/services": {
      "get": {
        "operationId": "listServices",
        "tags": [
          "services"
        ],
        "summary": "Services in an environment",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "env",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceList"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "syncloud_session"
      },
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Personal access token (syn_pat_\u2026)"
      },
      "accessKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "SYN1-HMAC-SHA256 Credential=<access key id>, Signature=<hex>, with X-Syn-Date and X-Syn-Content-Sha256 headers"
      }
    },
...
    "responses": {
      "Error": {
        "description": "Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ServiceSpec": {
        "type": "object",
        "required": [
          "image"
        ],
        "properties": {
          "image": {
            "type": "string",
            "example": "nginx:1.27"
          },
          "entrypoint": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "replaces the image ENTRYPOINT"
          },
          "command": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "env": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "sharedEnv": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "readOnly": true,
            "description": "snapshot of the environment's shared variables for this revision (set by the platform; env wins)"
          },
          "ports": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "container": {
                  "type": "integer"
                },
                "protocol": {
                  "type": "string",
                  "enum": [
                    "http",
                    "tcp",
                    "udp"
                  ]
                }
              }
            }
          },
          "resources": {
            "type": "object",
            "properties": {
              "cpu": {
                "type": "number",
                "description": "cores reserved (default 0.1)"
              },
              "memory": {
                "type": "integer",
                "description": "MiB reserved (default 128)"
              },
              "cpuLimit": {
                "type": "number"
              },
              "memoryLimit": {
                "type": "integer",
                "description": "MiB (default 2 \u00d7 memory)"
              }
            }
          },
          "placement": {
            "type": "object",
            "properties": {
              "strategy": {
                "type": "string",
                "enum": [
                  "spread",
                  "binpack"
                ]
              }
            }
          },
          "health": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "http",
                  "tcp",
                  "cmd"
                ]
              },
              "path": {
                "type": "string"
              },
              "port": {
                "type": "string"
              },
              "command": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "interval": {
                "type": "integer"
              },
              "timeout": {
                "type": "integer"
              },
              "retries": {
                "type": "integer"
              },
              "startPeriod": {
                "type": "integer"
              }
            }
          },
          "deployment": {
            "type": "object",
            "properties": {
              "circuitBreaker": {
                "type": "boolean",
                "default": true
              },
              "rollback": {
                "type": "boolean",
                "default": true
              }
            }
          }
        }
      },
...
      "Service": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "project": {
            "type": "string"
          },
          "environment": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "revision": {
            "type": "integer"
          },
          "desiredCount": {
            "type": "integer"
          },
          "running": {
            "type": "integer"
          },
          "pending": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "deleting": {
            "type": "boolean"
          },
          "spec": {
            "$ref": "#/components/schemas/ServiceSpec"
          },
          "endpoints": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "vip": {
            "type": "string",
            "description": "stable virtual IP inside the private network (\u00a78.6)"
          },
          "dnsName": {
            "type": "string",
            "example": "api.production.shop.syncloud.internal"
          },
          "deployment": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Deployment"
              }
            ],
            "nullable": true
          }
        }
      },
      "ServiceList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Service"
            }
          }
        }
      },
...
      "Deployment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "serviceId": {
            "type": "string"
          },
          "fromRevision": {
            "type": "integer"
          },
          "toRevision": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "in_progress",
              "succeeded",
              "failed",
              "rolled_back",
              "superseded",
              "waiting_hook"
            ]
          },
          "failedTasks": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
    }
  }
}